Added D3D11 libraries and arguments for HWA.
This change is basically a bit complicated approach using arguments to be able to load MKVs, MP4s at a much higher rates, allowing 60fps videos being able to be loaded in an instances, which are using shaders and lots of mods.
This change allows smooth 60fps videos (such as Numinous (https://www.youtube.com/watch?v=aht9ZSwpMCk)) to be loaded as intended, with all the fluid transitions and animations.
Also this change allows to properly support vp9/avc1 codec on certain machines and video cards, so they get not laggy when the video is played on runtime. (e.g. Cinematics mod allows to play videos, so when the cinematic starts on a laggy place, the cinematic will try to use hardware acceleration if possible to use more GPU, rather than CPU)

Feel free to throw tomatoes at me or make a PR to make this better, but I had literally no other idea what I can do here to make videos be able to be loaded just as smooth as I wanted to.